platforms: common calibration platform code
As some calibration code is common to all platforms, mainly general
shell commands, move this code into a common directory
"platforms/calibration/"
The platform name is now retrieved from MCUFIRMWARE_PLATFORM_BASE
makefile variable.
