+++ platforms/Makefile.include
@@ -4,4 +4,6 @@ DIRS += $(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)
 INCLUDES += -I$(MCUFIRMWAREBASE)/platforms/include/
 INCLUDES += -I$(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)/include
 
+CFLAGS += -DMCUFIRMWARE_PLATFORM_BASE=\"$(MCUFIRMWARE_PLATFORM_BASE)\"
+
 include $(MCUFIRMWAREBASE)/platforms/$(MCUFIRMWARE_PLATFORM_BASE)/Makefile.include
