+++ platforms/calibration/Makefile
@@ -1,0 +1,3 @@
+MODULE = calib_platform
+
+include $(RIOTBASE)/Makefile.base
